pyproject.toml
[project]
name = "chatbot-for-competitive-exams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-tavily",
    "tavily-python",
    "langgraph",
    "langchain-google-genai",
    "google-generativeai",
    "streamlit",
    "faiss-cpu",
    "sqlalchemy",
    "python-dotenv",
    "ipykernel",
    "fastapi",
    "uvicorn",
    "python-dotenv"
]
